Remove `cra` boilerplate files
These files should not live in the root. They don't even have any effect here.
Anything that's needed for the build should live in a `public/` folder.

These files are assets that are needed for the build.

On top of this, `asset-manigest.json` is a file generated by `cra` during the build.
As such, it shouldn't be placed in the `public/` folder.